+++ common/src/main/java/me/lucko/luckperms/common/commands/track/DeleteTrack.java
@@ -28,29 +28,31 @@ package me.lucko.luckperms.common.commands.track;
 import me.lucko.luckperms.common.actionlog.LoggedAction;
 import me.lucko.luckperms.common.command.CommandResult;
 import me.lucko.luckperms.common.command.abstraction.SingleCommand;
+import me.lucko.luckperms.common.command.access.ArgumentPermissions;
 import me.lucko.luckperms.common.command.access.CommandPermission;
+import me.lucko.luckperms.common.command.spec.CommandSpec;
 import me.lucko.luckperms.common.command.tabcomplete.TabCompleter;
 import me.lucko.luckperms.common.command.tabcomplete.TabCompletions;
-import me.lucko.luckperms.common.locale.LocaleManager;
-import me.lucko.luckperms.common.locale.command.CommandSpec;
-import me.lucko.luckperms.common.locale.message.Message;
+import me.lucko.luckperms.common.command.utils.ArgumentList;
+import me.lucko.luckperms.common.locale.Message;
 import me.lucko.luckperms.common.model.Track;
 import me.lucko.luckperms.common.plugin.LuckPermsPlugin;
 import me.lucko.luckperms.common.sender.Sender;
 import me.lucko.luckperms.common.util.Predicates;
 
+import net.kyori.adventure.text.Component;
 import net.luckperms.api.actionlog.Action;
 import net.luckperms.api.event.cause.DeletionCause;
 
 import java.util.List;
 
 public class DeleteTrack extends SingleCommand {
-    public DeleteTrack(LocaleManager locale) {
-        super(CommandSpec.DELETE_TRACK.localize(locale), "DeleteTrack", CommandPermission.DELETE_TRACK, Predicates.not(1));
+    public DeleteTrack() {
+        super(CommandSpec.DELETE_TRACK, "DeleteTrack", CommandPermission.DELETE_TRACK, Predicates.not(1));
     }
 
     @Override
-    public CommandResult execute(LuckPermsPlugin plugin, Sender sender, List<String> args, String label) {
+    public CommandResult execute(LuckPermsPlugin plugin, Sender sender, ArgumentList args, String label) {
         if (args.isEmpty()) {
             sendUsage(sender, label);
             return CommandResult.INVALID_ARGS;
@@ -63,15 +65,20 @@ public class DeleteTrack extends SingleCommand {
             return CommandResult.LOADING_ERROR;
         }
 
+        if (ArgumentPermissions.checkModifyPerms(plugin, sender, getPermission().get(), track)) {
+            Message.COMMAND_NO_PERMISSION.send(sender);
+            return CommandResult.NO_PERMISSION;
+        }
+
         try {
             plugin.getStorage().deleteTrack(track, DeletionCause.COMMAND).get();
         } catch (Exception e) {
-            e.printStackTrace();
-            Message.DELETE_ERROR.send(sender, track.getName());
+            plugin.getLogger().warn("Error whilst deleting track", e);
+            Message.DELETE_ERROR.send(sender, Component.text(track.getName()));
             return CommandResult.FAILURE;
         }
 
-        Message.DELETE_SUCCESS.send(sender, trackName);
+        Message.DELETE_SUCCESS.send(sender, Component.text(trackName));
 
         LoggedAction.build().source(sender).targetName(trackName).targetType(Action.Target.Type.TRACK)
                 .description("delete")
@@ -81,7 +88,7 @@ public class DeleteTrack extends SingleCommand {
     }
 
     @Override
-    public List<String> tabComplete(LuckPermsPlugin plugin, Sender sender, List<String> args) {
+    public List<String> tabComplete(LuckPermsPlugin plugin, Sender sender, ArgumentList args) {
         return TabCompleter.create()
                 .at(0, TabCompletions.tracks(plugin))
                 .complete(args);
